Fix compilation warnings.

+++ source/SAMRAI/algs/TimeRefinementIntegratorConnectorWidthRequestor.C
@@ -58,7 +58,7 @@ TimeRefinementIntegratorConnectorWidthRequestor::computeRequiredConnectorWidths(
    fine_connector_widths.resize(max_levels - 1, hier::IntVector::getZero(dim));
    self_connector_widths.clear();
    self_connector_widths.reserve(max_levels);
-   for ( size_t ln=0; ln<max_levels; ++ln ) {
+   for ( size_t ln=0; ln<static_cast<size_t>(max_levels); ++ln ) {
       hier::IntVector buffer(
          dim,
          d_tag_buffer.size() > ln ? d_tag_buffer[ln] : d_tag_buffer.back() );
